Move SUCCESS_REDIRECT_URL out of gmail-secret.json so can be more easily changed by others if required
