+++ server/routes/index.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const gmailAuth = require("./gmail-auth");
+const gmailAuth = require("./api/gmail-auth");
 
 router.get("/", (req, res, next) => {
   res.status(200).send("Home Page");
